requirements.txt
tensorflow
opencv-python
numpy
Flask-API